+++ example/lib/universal_ui/fake_ui.dart
@@ -1,0 +1,3 @@
+class PlatformViewRegistry {
+  static void registerViewFactory(String viewId, dynamic cb) {}
+}