+++ example/lib/universal_ui/real_ui.dart
@@ -1,0 +1,7 @@
+import 'dart:ui' as ui;
+
+class PlatformViewRegistry {
+  static void registerViewFactory(String viewId, dynamic cb) {
+    ui.platformViewRegistry.registerViewFactory(viewId, cb);
+  }
+}