+++ CHANGELOG.md
@@ -1,81 +1,251 @@
-## [0.0.1]
+## [1.9.1]
+* Cursor jumps to the most appropriate offset to display selection.
 
-* Rich text editor based on Quill Delta.
+## [1.9.0]
+* Support inline image.
 
-## [0.0.2]
-* Support image upload and launch url in read-only mode.
+## [1.8.3]
+* Updated quill_delta.
 
-## [0.0.3]
-* Update home page meta data.
+## [1.8.2]
+* Support mobile image alignment.
 
-## [0.0.4]
-* Update example.
+## [1.8.1]
+* Support mobile custom size image.
 
-## [0.0.5]
-* Update example.
+## [1.8.0]
+* Support entering link for image/video.
 
-## [0.0.6]
-* More toolbar functionality.
+## [1.7.3]
+* Bumps photo_view version.
 
-## [0.0.7]
-* Handle multiple image inserts.
-
-## [0.0.8]
-* Fix launching url.
-
-## [0.0.9]
-* Handle rgba color.
-
-## [0.1.0]
-* Fix insert image.
-
-## [0.1.1]
-* Fix cursor issue when undo.
-
-## [0.1.2]
-* Handle more text colors.
-
-## [0.1.3]
-* Handle cursor position change when undo/redo.
-
-## [0.1.4]
-* Handle url with trailing spaces.
-
-## [0.1.5]
-* Support text alignment.
-
-## [0.1.6]
-* Fix getExtentEndpointForSelection.
-
-## [0.1.7]
-* Support checked/unchecked list.
-
-## [0.1.8]
-* Support font and size attributes.
-
-## [0.2.0]
-* Add checked/unchecked list button in toolbar.
-
-## [0.2.1]
+## [1.7.2]
 * Fix static analysis error.
 
-## [0.2.2]
-* Update git repo.
+## [1.7.1]
+* Support Youtube video.
 
-## [0.2.3]
-* Support custom styles and image on local device storage without uploading.
+## [1.7.0]
+* Support video.
 
-## [0.2.4]
-* Support the use of custom icon size in toolbar.
+## [1.6.4]
+* Bug fix for clear format button.
 
-## [0.2.5]
-* Toggle text/background color button in toolbar.
+## [1.6.3]
+* Fixed dragging right handle scrolling issue.
 
-## [0.2.6]
-* Fix cursor after pasting.
+## [1.6.2]
+* Fixed the position of the selection status drag handle.
+
+## [1.6.1]
+* Upgrade image_picker and flutter_colorpicker.
+
+## [1.6.0]
+* Support Multi Row Toolbar.
+
+## [1.5.0]
+* Remove file_picker dependency.
+
+## [1.4.1]
+* Remove filesystem_picker dependency.
+
+## [1.4.0]
+* Remove path_provider dependency.
+
+## [1.3.4]
+* Add option to paintCursorAboveText.
+
+## [1.3.3]
+* Upgrade file_picker version.
+
+## [1.3.2]
+* Fix copy/paste bug.
+
+## [1.3.1]
+* New logo.
+
+## [1.3.0]
+* Support flutter 2.2.0.
+
+## [1.2.2]
+* Checkbox supports tapping.
+
+## [1.2.1]
+* Indented position not holding while editing.
+
+## [1.2.0]
+* Fix image button cancel causes crash.
+
+## [1.1.8]
+* Fix height of empty line bug.
+
+## [1.1.7]
+* Fix text selection in read-only mode.
+
+## [1.1.6]
+* Remove universal_html dependency.
+
+## [1.1.5]
+* Enable "Select", "Select All" and "Copy" in read-only mode.
+
+## [1.1.4]
+* Fix text selection issue.
+
+## [1.1.3]
+* Update example folder.
+
+## [1.1.2]
+* Add pedantic.
+
+## [1.1.1]
+* Base64 image support.
+
+## [1.1.0]
+* Support null safety.
+
+## [1.0.9]
+* Web support for raw editor and keyboard listener.
+
+## [1.0.8]
+* Support token attribute.
+
+## [1.0.7]
+* Fix crash on web (dart:io).
+
+## [1.0.6]
+* Add desktop support - WINDOWS, MACOS and LINUX.
+
+## [1.0.5]
+* Bug fix: Can not insert newline when Bold is toggled ON.
+
+## [1.0.4]
+* Upgrade photo_view to ^0.11.0.
+
+## [1.0.3]
+* Fix issue that text is not displayed while typing [WEB].
+
+## [1.0.2]
+* Update toolbar in sample home page.
+
+## [1.0.1]
+* Fix static analysis errors.
+
+## [1.0.0]
+* Support flutter 2.0.
+
+## [1.0.0-dev.2]
+* Improve link handling for tel, mailto and etc.
+
+## [1.0.0-dev.1]
+* Upgrade prerelease SDK & Bump for master.
+
+## [0.3.5]
+* Fix for cursor focus issues when keyboard is on.
+
+## [0.3.4]
+* Improve link handling for tel, mailto and etc.
+
+## [0.3.3]
+* More fix on cursor focus issue when keyboard is on.
+
+## [0.3.2]
+* Fix cursor focus issue when keyboard is on.
+
+## [0.3.1]
+* cursor focus when keyboard is on.
+
+## [0.3.0]
+* Line Height calculated based on font size.
+
+## [0.2.12]
+* Support placeholder.
+
+## [0.2.11]
+* Fix static analysis error.
+
+## [0.2.10]
+* Update TextInputConfiguration autocorrect to true in stable branch.
+
+## [0.2.9]
+* Update TextInputConfiguration autocorrect to true.
+
+## [0.2.8]
+* Support display local image besides network image in stable branch.
 
 ## [0.2.7]
 * Support display local image besides network image.
 
-## [0.2.8]
-* Support display local image besides network image in stable branch.
+## [0.2.6]
+* Fix cursor after pasting.
+
+## [0.2.5]
+* Toggle text/background color button in toolbar.
+
+## [0.2.4]
+* Support the use of custom icon size in toolbar.
+
+## [0.2.3]
+* Support custom styles and image on local device storage without uploading.
+
+## [0.2.2]
+* Update git repo.
+
+## [0.2.1]
+* Fix static analysis error.
+
+## [0.2.0]
+* Add checked/unchecked list button in toolbar.
+
+## [0.1.8]
+* Support font and size attributes.
+
+## [0.1.7]
+* Support checked/unchecked list.
+
+## [0.1.6]
+* Fix getExtentEndpointForSelection.
+
+## [0.1.5]
+* Support text alignment.
+
+## [0.1.4]
+* Handle url with trailing spaces.
+
+## [0.1.3]
+* Handle cursor position change when undo/redo.
+
+## [0.1.2]
+* Handle more text colors.
+
+## [0.1.1]
+* Fix cursor issue when undo.
+
+## [0.1.0]
+* Fix insert image.
+
+## [0.0.9]
+* Handle rgba color.
+
+## [0.0.8]
+* Fix launching url.
+
+## [0.0.7]
+* Handle multiple image inserts.
+
+## [0.0.6]
+* More toolbar functionality.
+
+## [0.0.5]
+* Update example.
+
+## [0.0.4]
+* Update example.
+
+## [0.0.3]
+* Update home page meta data.
+
+## [0.0.2]
+* Support image upload and launch url in read-only mode.
+
+## [0.0.1]
+* Rich text editor based on Quill Delta.